Complete clean architecture refactor with all fixes
Changes in this commit:
- Add missing GitHub token management commands
- Fix GitHub issues integration in context generation
- Fix GitHub issues loading indicator (spinner shows during fetch)
- Add DEVELOPMENT.md with concise architecture documentation
- Remove all deprecated code (extension-old, PromptTowerProvider, commands.ts, FileItem)

GitHub Integration Fixed:
- Issues now properly included in generated context
- Token counting works correctly with loading states
- Clean separation maintained (GitHub logic in services, not extension)

The extension now has:
- Full multi-workspace support
- Clean service-oriented architecture
- All original features working
- Much better maintainability
